Module cage::args

Data structures representing arguments that we can pass to docker-compose and other command-line tools.

Modules

opts
Command-line options which can be passed to docker-compose.

Structs

Command
A command which can be executed.

Enums

ActOn
The names of pods, services or both to pass to one of our commands.

Traits

ToArgs
Trait for types which can be converted to command-line arguments.