Module cage::cmd

Various commands which can be performed on a project, corresponding to CLI entry points.

To gain access to all commands at once:

use cage::cmd::*;Run

Traits

CommandCompose
Pass simple commands directly through to docker-compose.

CommandExec
We implement exec with a trait so we can put it in its own module.

CommandGenerate
Interface to various file-generation commands.

CommandPull
We implement pull with a trait so we put it in its own module.

CommandRepo
We implement repo with a trait so we put it in its own module.

CommandRun
We implement run with a trait so we put it in its own module.

CommandUp
We implement up with a trait so we put it in its own module.