Module cage::command_runner

Utilities for running and testing shell commands.

Structs

OsCommand
A wrapper around std::process:Command which logs the commands run.

OsCommandRunner
Support for running operating system commands.

TestCommand
A fake command that gets logged to a TestCommandRunner instead of actually getting run.

TestCommandRunner
Support for running commands in test mode.

Traits

Command
A stripped down interface based on std::process::Command. We use this so we can mock out shell commands during tests.

CommandRunner
A factory that produces objects conforming to our Command wrapper trait. During tests, we'll use this to mock out the underlying system and record all commands executed.