Trait cage::command_runner::Command

pub trait Command {
    fn arg<S: AsRef<OsStr>>(&mut self, arg: S) -> &mut Self;
    fn env<K, V>(&mut self, key: K, val: V) -> &mut Self where K: AsRef<OsStr>, V: AsRef<OsStr>;
    fn status(&mut self) -> Result<ExitStatus>;
    fn command_failed_error(&self) -> ErrorKind;

    fn args<S: AsRef<OsStr>>(&mut self, args: &[S]) -> &mut Self { ... }
    fn exec(&mut self) -> Result<()> { ... }
}

A stripped down interface based on std::process::Command. We use this so we can mock out shell commands during tests.

Required Methods

fn arg<S: AsRef<OsStr>>(&mut self, arg: S) -> &mut Self

Add an arugment to our command.

fn env<K, V>(&mut self, key: K, val: V) -> &mut Self where K: AsRef<OsStr>, V: AsRef<OsStr>

Set an environment variable for the process we're about to run.

fn status(&mut self) -> Result<ExitStatus>

Run our command.

fn command_failed_error(&self) -> ErrorKind

Make an error representing a failure of this command.

Provided Methods

fn args<S: AsRef<OsStr>>(&mut self, args: &[S]) -> &mut Self

Add several arguments to our command.

fn exec(&mut self) -> Result<()>

Run our command as per status, returning an error if the command fails.