Trait cage::command_runner::CommandRunner

pub trait CommandRunner {
    type Command: Command;
    fn build<S: AsRef<OsStr>>(&self, program: S) -> Self::Command;
}

A factory that produces objects conforming to our Command wrapper trait. During tests, we'll use this to mock out the underlying system and record all commands executed.

Associated Types

type Command: Command

The type of the commands we build. Must implement our custom Command trait.

Required Methods

fn build<S: AsRef<OsStr>>(&self, program: S) -> Self::Command

Build a new command.