Enum cage::ErrorKind

pub enum ErrorKind {
    Msg(String),
    Compose(ErrorKind),
    Utf8Error(FromUtf8Error),
    Glob(GlobError),
    GlobPattern(PatternError),
    Io(Error),
    StripPrefix(StripPrefixError),
    Vault(Error),
    CommandFailed(Vec<OsString>),
    CouldNotReadDirectory(PathBuf),
    CouldNotReadFile(PathBuf),
    CouldNotWriteFile(PathBuf),
    MismatchedVersion(VersionReq),
    PluginFailed(String),
    ServicesAddedInOverride(PathBuf, PathBuf, Vec<String>),
    UnknownPodOrService(String),
    UnknownService(String),
}

Variants
Msg(String)Compose(ErrorKind)Utf8Error(FromUtf8Error)Glob(GlobError)GlobPattern(PatternError)Io(Error)StripPrefix(StripPrefixError)Vault(Error)CommandFailed(Vec<OsString>)

An error occurred running an external command.

CouldNotReadDirectory(PathBuf)

An error occurred reading a directory.

CouldNotReadFile(PathBuf)

An error occurred reading a file.

CouldNotWriteFile(PathBuf)

An error occurred writing a file.

MismatchedVersion(VersionReq)

This project specified that it required a different version of this tool.

PluginFailed(String)

An error occurred applying a plugin.

ServicesAddedInOverride(PathBuf, PathBuf, Vec<String>)

An override file tried to add new services that weren't present in the file it was overriding.

UnknownPodOrService(String)

The requested pod or service does not appear to exist.

UnknownService(String)

The requested service does not appear to exist.

Methods

impl ErrorKind

fn description(&self) -> &str