Trait cage::plugins::Plugin

pub trait Plugin {
    fn name(&self) -> &'static str;
}

The "super-trait" of all our specific plugin traits. This needs to be usable as a trait object, so it may not contain any static "class" methods or methods with type parameters. Those can be found in PluginNew instead. This trait needs to be usable as a trait object so that we can create a Vec containing multiple different plugin implementations. Trait objects are as close as Rust comes to object orientation.

Required Methods

fn name(&self) -> &'static str

The name of this plugin (available after we create an instance).