Trait cage::plugins::PluginNew

pub trait PluginNew: Plugin + Sized + Debug {
    fn plugin_name() -> &'static str;
    fn new(project: &Project) -> Result<Self>;

    fn is_configured_for(_project: &Project) -> Result<bool> { ... }
}

Initialization for Plugin. These methods can't be part of Plugin itself, because they would prevent us from using Plugin as a trait object.

Required Methods

fn plugin_name() -> &'static str

The name of this plugin (available before we create an instance).

fn new(project: &Project) -> Result<Self>

Create a new plugin.

Provided Methods

fn is_configured_for(_project: &Project) -> Result<bool>

Has this plugin been configured for this project? This will be called before instantiating any plugin type except PluginGenerate.