Struct cage::Pod

pub struct Pod { /* fields omitted */ }

A pod, specified by pods/$NAME.yml and zero or more pods/overrides/*/*.yml overrides that we can apply to it.

Methods

impl Pod

fn name(&self) -> &str

Get the name of this pod.

fn pod_type(&self) -> PodType

Get the type of this pod.

fn service_names(&self) -> &BTreeSet<String>

Get the names of the services declared in this pod.

fn enabled_in(&self, ovr: &Override) -> bool

Is this pod enabled in the specified override?

fn base_dir(&self) -> &Path

The base directory for our relative paths.

fn rel_path(&self) -> &Path

The path to the top-level file defining this pod, relative to the base_dir specified at creation time.

fn file(&self) -> &File

The top-level file defining this pod. This is normalized to include the appropriate env_file entries, but if you want to do more complicated transformations, you'll need to clone it with to_owned() first.

fn override_rel_path(&self, ovr: &Override) -> Result<&Path>

The path to the specificied override file for this pod.

fn override_file(&self, ovr: &Override) -> Result<&File>

The dc::File for this override.

fn merged_file(&self, ovr: &Override) -> Result<File>

Return the base file and the override file merged into a single docker-compose.yml file.

fn override_files(&self) -> OverrideFiles

All the overrides associated with this pod.

fn all_files(&self) -> AllFiles

Iterate over all dc::File objects associated with this pod, including both the main file() and all the files in override_files().

fn service(&self, ovr: &Override, name: &str) -> Result<Option<Service>>

Look up a service by name.

fn service_or_err(&self, ovr: &Override, name: &str) -> Result<Service>

Like service, but returns an error if the service can't be found.

fn compose_args(&self, proj: &Project, ovr: &Override) -> Result<Vec<OsString>>

Command-line -p and -f arguments that we'll pass to docker-compose to describe this file.