Struct cage::Project

pub struct Project { /* fields omitted */ }

A cage project, which is represented as a directory containing a pods subdirectory.

Methods

impl Project

fn from_current_dir() -> Result<Project>

Create a Project using the pre-existing project files in the current directory as input and the .cage subdirectory as output.

use cage::Project;
use std::env;

let saved = env::current_dir().unwrap();
env::set_current_dir("examples/hello/pods").unwrap();

let proj = Project::from_current_dir().unwrap();
assert_eq!(proj.root_dir(), saved.join("examples/hello"));
assert_eq!(proj.src_dir(), saved.join("examples/hello/src"));
assert_eq!(proj.output_dir(), saved.join("examples/hello/.cage"));

env::set_current_dir(saved).unwrap();Run

fn name(&self) -> &str

The name of this project. This defaults to the name of the current directory.

fn set_name(&mut self, name: &str) -> &mut Project

Set the name of this project. This should be done before calling output or any methods in cmd.

fn root_dir(&self) -> &Path

The root directory of this project.

fn src_dir(&self) -> &Path

The source directory of this project, where we can put cloned git repositories.

fn output_dir(&self) -> &Path

The output directory of this project. Normally .cage inside the root_dir, but it may be overriden.

fn pods_dir(&self) -> PathBuf

The directory in which are pods are defined, and relative to which all docker-compose.yml paths should be interpreted.

fn output_pods_dir(&self) -> PathBuf

The path relative to which our pods will be output. This can be joined with Pod::rel_path to get an output path for a specific pod.

fn pods(&self) -> Pods

Iterate over all pods in this project.

fn pod(&self, name: &str) -> Option<&Pod>

Look up the named pod.

fn service<'a>(&self, name: &'a str) -> Option<(&Pod, &str)>

Look up the named service. Returns the pod containing the service and the name of the service within that pod.

fn service_or_err<'a>(&self, name: &'a str) -> Result<(&Pod, &str)>

Like service, but returns an error if the service is unknown.

fn pod_or_service<'a, 'b>(&'a self, name: &'b str) -> Option<PodOrService<'a>>

Look for a name as a pod first, and if that fails, look for it as a service.

fn pod_or_service_or_err<'a, 'b>(&'a self, name: &'b str) -> Result<PodOrService<'a>>

Like pod_or_service, but returns an error if no pod or service of that name can be found.

fn overrides(&self) -> Overrides

Iterate over all overrides in this project.

fn ovr(&self, name: &str) -> Option<&Override>

Look up the named override. We name this function ovr instead of override to avoid a keyword clash.

fn repos(&self) -> &Repos

Return the collection of git repositories associated with this project.

fn hooks(&self) -> &HookManager

Get our available hooks.

fn default_tags(&self) -> Option<&DefaultTags>

Get the default tags associated with this project, if any.

fn set_default_tags(&mut self, tags: DefaultTags) -> &mut Project

Set the default tags associated with this project.

fn plugins(&self) -> &Manager

Our plugin manager.

fn output(&self, ovr: &Override) -> Result<()>

Delete our existing output and replace it with a processed and expanded version of our pod definitions.

fn export(&self, ovr: &Override, export_dir: &Path) -> Result<()>

Export this project (with the specified override applied) as a set of standalone *.yml files with no environment variable interpolations and no external dependencies.