Struct cage::ProjectConfig

pub struct ProjectConfig {
    pub cage_version: Option<VersionReq>,
    // some fields omitted
}

Configuration information about a project, read in from PROJECT_CONFIG_PATH.

Fields
cage_version: Option<VersionReq>

A semantic version requirement specifying compatible versions of this tool.

Methods

impl ProjectConfig

fn new(path: &Path) -> Result<Self>

Load a config file from the specified path.