Struct cage::Repo

pub struct Repo { /* fields omitted */ }

A single repository.

Methods

impl Repo

fn alias(&self) -> &str

A short local name for this git repository, suitable for use as a directory name or command-line argument.

fn git_url(&self) -> &GitUrl

The remote git URL from which we can clone this repository.

fn rel_path(&self) -> PathBuf

The path to which we would check out this repository, relative to Project::src_dir.

fn path(&self, project: &Project) -> PathBuf

The full path to which we would check out this repository. The project argument is mandatory because we can't store a pointer to it without creating a circular reference loop.

fn is_cloned(&self, project: &Project) -> bool

Has this project been cloned locally?

fn clone_source<CR>(&self, runner: &CR, project: &Project) -> Result<()> where CR: CommandRunner

Clone the source code of this repository using git.